refactor: 캘린더 설정 페이지 - 컴포넌트 이름 수정 및 path 수정
